feat: implementação
CRUD: Usuários
CRUD: Automóveis
New pages layout

Falta coisa ainda(bastante)

+++ backend/src/main/java/com/aluguel/models/Rendimento.java
@@ -2,19 +2,20 @@ package com.aluguel.models;
 
 import io.micronaut.serde.annotation.Serdeable;
 import jakarta.persistence.Embeddable;
+import java.math.BigDecimal;
 
 @Serdeable
-@Embeddable // Diz que esses dados vão ser embutidos na tabela auxiliar de rendimentos
+@Embeddable
 public class Rendimento {
-
-    private String empregadora;
-    private Double valor;
+    // Usando o termo exato do PDF para evitar erro na Facade
+    private String empregadora; 
+    private BigDecimal valor;
 
     public Rendimento() {}
 
     public String getEmpregadora() { return empregadora; }
     public void setEmpregadora(String empregadora) { this.empregadora = empregadora; }
 
-    public Double getValor() { return valor; }
-    public void setValor(Double valor) { this.valor = valor; }
+    public BigDecimal getValor() { return valor; }
+    public void setValor(BigDecimal valor) { this.valor = valor; }
 }
